Project Euler Problem 03 Update

+++ problem03.py
@@ -1,7 +1,7 @@
 import math
 
 def primeFactor(n):
-	for i in xrange(2,math.sqrt(n)+1):
+	for i in xrange(2,int(math.sqrt(n))+1):
 		if i != n and n % i == 0:
 			n = n / i
 			i = i - 1
